Remove extra casting

+++ src/io/kson_io.cpp
@@ -45,7 +45,7 @@ namespace
 		else if constexpr (std::is_floating_point_v<T>)
 		{
 			// To avoid precision problems (e.g. "0.7000000000000001"), round the value to eight decimal places
-			json.emplace(key, static_cast<std::int64_t>(std::round(value * 1e8)) / 1e8);
+			json.emplace(key, std::round(value * 1e8) / 1e8);
 		}
 		else if constexpr (std::is_same_v<T, nlohmann::json>)
 		{
